[Change] Models separation.
