+++ virtualenv/SaaS/api/models/skill.py
@@ -1,0 +1,9 @@
+from django.db import models
+
+
+class Skill(models.Model):
+    id = models.AutoField(primary_key=True)
+    name = models.CharField(max_length=100)
+
+    class Meta:
+        db_table = "Skill"
